+++ docs/ci.puml
@@ -4,25 +4,26 @@
 namespace numopt {
 
 
-class DavidsonSolver {
-    __ private members __
-        - const double residue_tolerance
-        - const double correction_threshold
-        - const size_t maximum_subspace_dimension
+    class DavidsonSolver {
+        __ private members __
+            - const double residue_tolerance
+            - const double correction_threshold
+            - const size_t maximum_subspace_dimension
 
-        - const Eigen::VectorXd t0
-        - const numopt::VectorFunction matrixVectorProduct
+            - const Eigen::VectorXd t0
+            - const numopt::VectorFunction matrixVectorProduct
 
-        - double eigenvalue
-        - Eigen::VectorXd eigenvector
-    == constructors ==
-        **DavidsonSolver**(const numopt::VectorFunction& matrixVectorProduct, const Eigen::VectorXd& t0, double residue_tolerance = 1.0e-08, double correction_threshold = 1.0e-03, size_t maximum_subspace_dimension = 15)
-    __ public methods __
-        + **get_eigenvalue**()
-        + **get_eigenvector**()
+            - double eigenvalue
+            - Eigen::VectorXd eigenvector
+        == constructors ==
+            **DavidsonSolver**(const numopt::VectorFunction& matrixVectorProduct, const Eigen::VectorXd& t_0, const Eigen::VectorXd& diagonal, double residue_tolerance = 1.0e-08, double correction_threshold = 1.0e-03, size_t maximum_subspace_dimension = 15)
+            **DavidsonSolver**(const Eigen::MatrixXd& A, const Eigen::VectorXd& t_0, double residue_tolerance = 1.0e-08, double correction_threshold = 1.0e-03, size_t maximum_subspace_dimension = 15);
+        __ public methods __
+            + **get_eigenvalue**()
+            + **get_eigenvector**()
 
-        + **solve**()
-}
+            + **solve**()
+    }
 
 
 }
