+++ include/BaseSolver.hpp
@@ -1,0 +1,10 @@
+#ifndef CI_BASESOLVER_HPP
+#define CI_BASESOLVER_HPP
+
+
+class BaseSolver {
+
+};
+
+
+#endif  // CI_BASESOLVER_HPP
