+++ include/DavidsonSolver.hpp
@@ -1,0 +1,10 @@
+#ifndef CI_DAVIDSONSOLVER_HPP
+#define CI_DAVIDSONSOLVER_HPP
+
+
+class DavidsonSolver {
+
+};
+
+
+#endif  // CI_DAVIDSONSOLVER_HPP
