+++ include/DenseSolver.hpp
@@ -1,0 +1,10 @@
+#ifndef CI_DENSESOLVER_HPP
+#define CI_DENSESOLVER_HPP
+
+
+class DenseSolver {
+
+};
+
+
+#endif  // CI_DENSESOLVER_HPP
