+++ include/FCI.hpp
@@ -1,0 +1,10 @@
+#ifndef CI_FCI_HPP
+#define CI_FCI_HPP
+
+
+class FCI {
+
+};
+
+
+#endif  // CI_FCI_HPP
