+++ include/MatrixSolver.hpp
@@ -1,0 +1,10 @@
+#ifndef CI_MATRIXSOLVER_HPP
+#define CI_MATRIXSOLVER_HPP
+
+
+class MatrixSolver {
+
+};
+
+
+#endif  // CI_MATRIXSOLVER_HPP
