+++ include/SparseSolver.hpp
@@ -1,0 +1,10 @@
+#ifndef CI_SPARSESOLVER_HPP
+#define CI_SPARSESOLVER_HPP
+
+
+class SparseSolver {
+
+};
+
+
+#endif  // CI_SPARSESOLVER_HPP
