+++ not_yet_parsed/BaseHamiltonianSolver.cpp
