+++ not_yet_parsed/BaseHamiltonianSolver.hpp
