+++ not_yet_parsed/DenseHamiltonianSolver.hpp
